wip - sign and fix

+++ trunk/LibZ.Tool/Tasks/TaskBase.cs
@@ -2,7 +2,9 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Reflection;
 using System.Security.Cryptography;
+using System.Security.Cryptography.X509Certificates;
 using System.Text;
 using Mono.Cecil;
 
@@ -20,12 +22,14 @@ namespace LibZ.Tool.Tasks
 
 		protected static void DeleteFile(string fileName)
 		{
+			if (!File.Exists(fileName)) return;
+
 			try
 			{
 				Log.Debug("Deleting '{0}'", fileName);
 				File.Delete(fileName);
 			}
-				// ReSharper disable EmptyGeneralCatchClause
+			// ReSharper disable EmptyGeneralCatchClause
 			catch
 			{
 				Log.Warn("File '{0}' could not be deleted", fileName);
@@ -50,6 +54,11 @@ namespace LibZ.Tool.Tasks
 
 		#region reflection utilities
 
+		protected static bool EqualAssemblyNames(string valueA, string valueB)
+		{
+			return string.Compare(valueA, valueB, StringComparison.InvariantCultureIgnoreCase) == 0;
+		}
+
 		protected static string GetAssemblyName(AssemblyDefinition assembly)
 		{
 			return assembly.Name.FullName;
@@ -72,6 +81,90 @@ namespace LibZ.Tool.Tasks
 			return AssemblyArchitecture.AnyCPU;
 		}
 
+		protected static bool IsSigned(AssemblyDefinition assembly)
+		{
+			return assembly.Modules.Any(m => (m.Attributes & ModuleAttributes.StrongNameSigned) != 0);
+		}
+
+		protected static StrongNameKeyPair LoadKeyPair(string keyFileName, string password)
+		{
+			if (String.IsNullOrWhiteSpace(keyFileName)) return null;
+
+			Log.Info("Loading singing key from '{0}'", keyFileName);
+			// do not use constructor with filename it does not really load the key (?)
+
+			var keyPair =
+				string.IsNullOrWhiteSpace(password)
+				? new StrongNameKeyPair(File.ReadAllBytes(keyFileName))
+				: GetStrongNameKeyPairFromPfx(keyFileName, password);
+
+			try
+			{
+				var publicKey = keyPair.PublicKey;
+				// this is not important, just wanted to clutter screen a little bit
+				// there is no built-in ToHexString, and as it is not really needed 
+				// I won't write it so I use ToBase64String
+				Log.Debug("Public key is '{0}'", Convert.ToBase64String(publicKey));
+			}
+			catch
+			{
+				Log.Error("There is a chance this is kind of well-known problem");
+				Log.Error(
+					"See 'http://stackoverflow.com/questions/5659740/unable-to-obtain-public-key-for-strongnamekeypair' for details");
+				throw;
+			}
+			return keyPair;
+		}
+
+		protected static StrongNameKeyPair GetStrongNameKeyPairFromPfx(string pfxFile, string password)
+		{
+			// http://stackoverflow.com/questions/7556846/how-to-use-strongnamekeypair-with-a-password-protected-keyfile-pfx
+
+			var certs = new X509Certificate2Collection();
+			certs.Import(pfxFile, password, X509KeyStorageFlags.Exportable);
+			if (certs.Count == 0)
+				throw new ArgumentException(null, "pfxFile");
+
+			var provider = certs[0].PrivateKey as RSACryptoServiceProvider;
+			if (provider == null) // not a good pfx file
+				throw new ArgumentException(null, "pfxFile");
+
+			return new StrongNameKeyPair(provider.ExportCspBlob(false));
+		}
+
+		protected static AssemblyDefinition LoadAssembly(string assemblyFileName)
+		{
+			Log.Debug("Loading '{0}'", assemblyFileName);
+			return AssemblyDefinition.ReadAssembly(assemblyFileName);
+		}
+
+		protected void SaveAssembly(AssemblyDefinition assembly, string assemblyFileName, StrongNameKeyPair keyPair = null)
+		{
+			var tempFileName = assemblyFileName + ".temp";
+
+			try
+			{
+				if (keyPair == null)
+				{
+					Log.Debug("Saving '{0}'", assemblyFileName);
+					assembly.Write(tempFileName);
+				}
+				else
+				{
+					Log.Debug("Saving and signing '{0}'", assemblyFileName);
+					assembly.Write(tempFileName, new WriterParameters { StrongNameKeyPair = keyPair });
+				}
+
+				File.Delete(assemblyFileName);
+				File.Move(tempFileName, assemblyFileName);
+			}
+			catch
+			{
+				if (File.Exists(tempFileName)) DeleteFile(tempFileName);
+				throw;
+			}
+		}
+
 		#endregion
 
 		#region exceptions
@@ -83,7 +176,7 @@ namespace LibZ.Tool.Tasks
 
 		protected static Exception FileNotFound(string fileName)
 		{
-			return new FileNotFoundException(string.Format("File '{0}' could not be found", fileName));
+			return new FileNotFoundException(String.Format("File '{0}' could not be found", fileName));
 		}
 
 		#endregion
